refactor: Category, ImageSlider, SearchBar 컴포넌트를 mainPage 폴더로 이동
