+++ src/components/mainPage/ImageSlider.jsx
